Fix double-promotion the other way around
Stay close to the data format in absence of calculations

+++ KappaAnalysis/interface/Filters/GoodPrimaryVertexFilter.h
@@ -24,7 +24,7 @@ public:
 
 private:
 
-	double m_maxPrimaryVertexZ;
-	double m_maxPrimaryVertexRho;
-	double m_minPrimaryVertexFitnDOF;
+	float m_maxPrimaryVertexZ;
+	float m_maxPrimaryVertexRho;
+	float m_minPrimaryVertexFitnDOF;
 };
