+++ KappaAnalysis/src/Filters/GoodPrimaryVertexFilter.cc
@@ -23,9 +23,9 @@ bool GoodPrimaryVertexFilter::DoesEventPass(KappaEvent const& event, KappaProduc
 	assert(event.m_vertexSummary);
 	bool pass = false;
 	if (!(event.m_vertexSummary->pv.fake()) &&
-	   (static_cast<double>(std::abs(event.m_vertexSummary->pv.position.Z())) < m_maxPrimaryVertexZ) &&
-	   (static_cast<double>(std::abs(event.m_vertexSummary->pv.position.Rho())) < m_maxPrimaryVertexRho) &&
-	   (static_cast<double>(event.m_vertexSummary->pv.nDOF) > m_minPrimaryVertexFitnDOF))
+	   (std::abs(event.m_vertexSummary->pv.position.Z()) < m_maxPrimaryVertexZ) &&
+	   (std::abs(event.m_vertexSummary->pv.position.Rho()) < m_maxPrimaryVertexRho) &&
+	   (event.m_vertexSummary->pv.nDOF > m_minPrimaryVertexFitnDOF))
 	{
 		pass = true;
 	}
